Move surface transformer/lator to detail
