+++ src/orange/CMakeLists.txt
@@ -28,8 +28,8 @@ list(APPEND SOURCES
   surf/SimpleQuadric.cc
   surf/Sphere.cc
   surf/SurfaceIO.cc
-  surf/SurfaceTranslator.cc
-  surf/SurfaceTransformer.cc
+  surf/detail/SurfaceTranslator.cc
+  surf/detail/SurfaceTransformer.cc
   transform/Transformation.cc
 )
