+++ test/CMakeLists.txt
@@ -229,8 +229,8 @@ celeritas_add_test(orange/surf/PlaneAligned.test.cc)
 celeritas_add_test(orange/surf/SimpleQuadric.test.cc)
 celeritas_add_test(orange/surf/Sphere.test.cc)
 celeritas_add_test(orange/surf/SphereCentered.test.cc)
-celeritas_add_test(orange/surf/SurfaceTranslator.test.cc)
-celeritas_add_test(orange/surf/SurfaceTransformer.test.cc)
+celeritas_add_test(orange/surf/detail/SurfaceTranslator.test.cc)
+celeritas_add_test(orange/surf/detail/SurfaceTransformer.test.cc)
 celeritas_add_device_test(orange/surf/SurfaceAction)
 
 #-------------------------------------#
